refactor: remove Graph View component
